fix(review): centralize artifact retirement ownership
