+++ tests/contracts/test_plugin_source_export_chain.py
@@ -40,6 +40,7 @@ ADDED = (
     "PluginArtifactKind",
     "PluginArtifactIdentity",
     "PluginArtifactPublicationError",
+    "PluginArtifactRetirementEngine",
     "PluginArtifactRetirementOwner",
     "PluginArtifactValidationError",
     "PluginLaunchBinding",
